Collect & transfer, sketch.
